requirements.txt
fastapi
uvicorn
python-multipart
onnxruntime
opencv-python-headless
Pillow
numpy
rembg
huggingface_hub
